backend/requirements.txt
boto3==1.34.8
celery[redis]==5.2.7
cryptography==41.0.7
django-cors-headers==3.10.0
django-crum==0.7.9
django-environ==0.8.1
django-extensions==3.2.1
django-filter==21.1
django-fsm-admin==1.2.4
django-fsm==2.8.0
django-model-utils==4.2.0
django-redis
django-sortedm2m==3.1.1
django-storages==1.13.2
django==4.1.13
djangorestframework-simplejwt==5.0.0
djangorestframework==3.12.4
django-extensions==3.2.1
fcm-django==2.0.0
ffmpeg-python==0.2.0
firebase-admin==6.1.0
google-api-python-client==2.36
gunicorn==20.1.0
inapppy==2.5.2
ipython==8.11.0
locust
markdown==3.3.4
mutagen
opencv-python==4.5.4.58
Pillow==10.1.0
psycopg2-binary==2.9.9
redis==4.5.4
requests>=2.28.1,<2.29
sentry-sdk==1.20.0
shortuuid==1.0.11
stringcase==1.2.0
toolz==0.12.0
twilio==7.5.1
